pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grade-na-hora-utfpr"
version = "0.1.0"
description = "App desktop para montar grade horaria da UTFPR"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "customtkinter>=5.2",
  "playwright>=1.50",
  "Pillow>=10.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.9",
]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
addopts = "-q -ra"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]